pyprojection: fix size comparisons for 32-bit builds

+++ projection/winrt-Windows.ApplicationModel.Resources.Core/py.Windows.ApplicationModel.Resources.Core.cpp
@@ -783,7 +783,7 @@ namespace py::cpp::Windows::ApplicationModel::Resources::Core
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -1632,7 +1632,7 @@ namespace py::cpp::Windows::ApplicationModel::Resources::Core
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -4325,7 +4325,7 @@ namespace py::cpp::Windows::ApplicationModel::Resources::Core
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
