+++ projection/winrt-Windows.Data.Xml.Dom/py.Windows.Data.Xml.Dom.cpp
@@ -11260,7 +11260,7 @@ namespace py::cpp::Windows::Data::Xml::Dom
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -11660,7 +11660,7 @@ namespace py::cpp::Windows::Data::Xml::Dom
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
