+++ projection/winrt-Windows.Networking.ServiceDiscovery.Dnssd/py.Windows.Networking.ServiceDiscovery.Dnssd.cpp
@@ -993,7 +993,7 @@ namespace py::cpp::Windows::Networking::ServiceDiscovery::Dnssd
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
