+++ projection/winrt-Windows.Storage.Pickers/py.Windows.Storage.Pickers.cpp
@@ -544,7 +544,7 @@ namespace py::cpp::Windows::Storage::Pickers
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -1864,7 +1864,7 @@ namespace py::cpp::Windows::Storage::Pickers
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
