+++ projection/winrt-Windows.UI.Xaml.Data/py.Windows.UI.Xaml.Data.cpp
@@ -2954,7 +2954,7 @@ namespace py::cpp::Windows::UI::Xaml::Data
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
