+++ projection/winrt-Windows.UI.Xaml.Media/py.Windows.UI.Xaml.Media.cpp
@@ -2303,7 +2303,7 @@ namespace py::cpp::Windows::UI::Xaml::Media
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -3938,7 +3938,7 @@ namespace py::cpp::Windows::UI::Xaml::Media
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -5557,7 +5557,7 @@ namespace py::cpp::Windows::UI::Xaml::Media
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -7051,7 +7051,7 @@ namespace py::cpp::Windows::UI::Xaml::Media
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -10353,7 +10353,7 @@ namespace py::cpp::Windows::UI::Xaml::Media
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -11308,7 +11308,7 @@ namespace py::cpp::Windows::UI::Xaml::Media
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -12914,7 +12914,7 @@ namespace py::cpp::Windows::UI::Xaml::Media
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -17478,7 +17478,7 @@ namespace py::cpp::Windows::UI::Xaml::Media
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -18325,7 +18325,7 @@ namespace py::cpp::Windows::UI::Xaml::Media
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
