+++ projection/winrt-Windows.Web.Http.Headers/py.Windows.Web.Http.Headers.cpp
@@ -837,7 +837,7 @@ namespace py::cpp::Windows::Web::Http::Headers
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -1857,7 +1857,7 @@ namespace py::cpp::Windows::Web::Http::Headers
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -2817,7 +2817,7 @@ namespace py::cpp::Windows::Web::Http::Headers
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -3777,7 +3777,7 @@ namespace py::cpp::Windows::Web::Http::Headers
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -4773,7 +4773,7 @@ namespace py::cpp::Windows::Web::Http::Headers
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -7590,7 +7590,7 @@ namespace py::cpp::Windows::Web::Http::Headers
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -9154,7 +9154,7 @@ namespace py::cpp::Windows::Web::Http::Headers
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -9880,7 +9880,7 @@ namespace py::cpp::Windows::Web::Http::Headers
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -10876,7 +10876,7 @@ namespace py::cpp::Windows::Web::Http::Headers
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -12326,7 +12326,7 @@ namespace py::cpp::Windows::Web::Http::Headers
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -13052,7 +13052,7 @@ namespace py::cpp::Windows::Web::Http::Headers
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -14616,7 +14616,7 @@ namespace py::cpp::Windows::Web::Http::Headers
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -17556,7 +17556,7 @@ namespace py::cpp::Windows::Web::Http::Headers
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
